Core/AI: EventMap groups and phases were lost in a uint32
ScheduleEvent OR-ed the group (bits 32-47) and phase (bits 48-63) flags into
the uint32 event id, so they were always dropped: CancelEventGroup, grouped
DelayEvents and the phase filter of ExecuteEvent never did anything since the
event map rewrite. Build the 64-bit value instead, and accept group and phase
16 like the rest of the API. GetPhaseMask also truncated the uint16 mask.

+++ src/server/game/AI/EventMap.h
@@ -144,7 +144,7 @@ public:
     * @name GetPhaseMask
     * @return Active phases as mask.
     */
-    uint8 GetPhaseMask() const
+    uint16 GetPhaseMask() const
     {
         return _phase;
     }
